Overhaul profile area in part.
* Remove character stats
* Fold character posts and topics under characters
* Tidy button navs
