+++ Themes/natural/css/_profile.scss
@@ -1,0 +1,80 @@
+.profile {
+	&.cat_bar {
+		background: $tertiary-accent;
+	}
+
+	&.profile-summary {
+		display: flex;
+		flex-flow: row wrap;
+
+		.user-summary {
+			flex-basis: calc(#{$avatar_max_width}px + 100px);
+			padding-right: 30px;
+		}
+
+		.profile-online {
+			text-align: center;
+		}
+
+		.user-actions {
+			margin-top: 10px;
+			font-size: 30px;
+
+			text-align: center;
+
+			a {
+				display: inline-block;
+			}
+		}
+
+		.user-details {
+			h5 {
+				margin-top: 20px;
+				margin-bottom: 10px;
+			}
+			.characters {
+				li {
+					display: table-row;
+				}
+
+				.character-avatar {
+					display: table-cell;
+					width: 50px;
+					height: 50px;
+					text-align: center;
+					vertical-align: middle;
+
+					img {
+						max-width: 50px;
+						max-height: 50px;
+					}
+				}
+
+				.character-name {
+					display: table-cell;
+					vertical-align: middle;
+					padding-left: 30px;
+				}
+			}
+		}
+
+		dl {
+			display: grid;
+			grid-template: auto / 15em 1fr;
+			row-gap: 0.5em;
+
+			dt {
+				grid-column: 1;
+			}
+
+			dd {
+				grid-column: 2;
+			}
+
+			dt.full-width {
+				grid-column-start: 1;
+				grid-column-end: 3;
+			}
+		}
+	}
+}
