+++ Themes/natural/css/index.scss
@@ -28,6 +28,7 @@
 /* Specific functions and pages */
 @import "natural/_character_list.scss";
 @import "natural/_contactform.scss";
+@import "natural/_profile.scss";
 @import "natural/_search.scss";
 @import "natural/_help.scss";
 @import "natural/_shipper.scss";
