+++ Themes/natural/partials/sidebar_navigation_above.hbs
@@ -26,7 +26,7 @@
 						{{/if}}
 						<ul>
 							{{#each links}}
-								<li{{#if active}} class="active"{{/if}}><a href="{{url}}">{{label}}</a></li>
+								<li{{#if active}} class="active"{{/if}}><a href="{{url}}">{{{label}}}</a></li>
 							{{/each}}
 						</ul>
 					{{/each}}
