+++ Themes/natural/templates/profile_character_posts.hbs
@@ -1,7 +1,7 @@
 		{{! Showing a character's posts/topics }}
-		<div class="cat_bar">
+		<div class="cat_bar profile">
 			<h3 class="catbg">
-				{{#if context.is_topics}}{{{txt.showTopics}}}{{else}}{{{txt.showMessages}}}{{/if}} - {{{context.character.character_name}}}
+				{{{context.character.character_name}}} - {{#if context.is_topics}}{{{txt.showTopics}}}{{else}}{{{txt.showMessages}}}{{/if}}
 			</h3>
 		</div>
