+++ Themes/natural/templates/profile_character_sheet.hbs
@@ -1,7 +1,7 @@
 			{{! Showing a character sheet }}
-			<div class="cat_bar">
+			<div class="cat_bar profile">
 				<h3 class="catbg">
-					{{{txt.char_sheet}}} - {{{context.character.character_name}}}
+					{{{context.character.character_name}}} - {{{txt.char_sheet}}}
 				</h3>
 			</div>
