+++ Themes/natural/templates/profile_character_sheet_history.hbs
@@ -1,8 +1,13 @@
 		{{! History of character sheets. }}
-		<div class="cat_bar">
+		<div class="cat_bar profile">
 			<h3 class="catbg">
-				{{{txt.char_sheet}}} - {{{context.character.character_name}}} - {{{txt.char_sheet_history}}}
+				{{{context.character.character_name}}} - {{{txt.char_sheet_history_title}}}
 			</h3>
+		</div>
+
+		<div class="buttonlist">
+			<a href="{{scripturl}}?action=profile;u={{context.id_member}};area=characters;char={{context.character.id_character}}" class="button">{{{context.character.character_name}}}</a>
+			<a href="{{scripturl}}?action=profile;u={{context.id_member}};area=character_sheet;char={{context.character.id_character}}" class="button">{{{txt.char_sheet_link}}}</a>
 		</div>
 
 {{#each context.history_items}}
